'feature' config option
This introduces a 'features' option which currently contains
3 AppArmor related options: netmount, blockmount and
nesting. These change the container's AppArmor profile and
are thus incompatible with a custom 'lxc.aa_profile' option,
and are used to allow mounting of network or block devices
as well as nested containers. (None of them are recommended
as they come with several security concerns.)

There's now a new /lxc/$vmid/features API call which returns
a list of available features with their descriptions and
whether the current user is allowed to use them. (Currently
all 3 features require being logged in as root@pam to be
activated, but no special permissions to be _deactivated_.)
Next to querying this API call also allows modifying the
features, but this can also be done via the regular
/lxc/vmid/config API call.

At build-time the AppArmor profiles for all combinations of
features are produced and installed as
/etc/apparmor.d/lxc/lxc-pve-profiles.

Since each feature can have its own 'permissions' and are
thus compared to the old setting when updating the config
the check_ct_modify_config_perm() now gets the old config
passed as parameter for more flexibility.